backend/requirements.txt
livekit-agents[openai,silero,deepgram]
aiohttp
python-dotenv
